request added

+++ src/main/java/com/weather/models/iqair/Weather.java
@@ -9,10 +9,10 @@ import java.util.Date;
 @ToString
 public class Weather{
     private Date ts;
-    private int tp;
-    private int pr;
-    private int hu;
+    private float tp;
+    private float pr;
+    private float hu;
     private double ws;
-    private int wd;
+    private float wd;
     private String ic;
 }
